% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bdesert_stack}
\alias{Bdesert_stack}
\title{Blooming Desert Northern Chile}
\format{
RasterStack EPSG:32719
}
\source{
\url{https://lpdaac.usgs.gov/}
}
\usage{
Bdesert_stack
}
\description{
MODIS NDVI RasterStack of the Atacama Desert, Northern Chile, combining MOD13Q1 & MYD13Q1 between 2000.02.18 and 2021.06.26,
cleaned using the detailed QA band. Blooming deserts are rare greening events to be seen as positive NDVI anomalies in Atacama
}
\keyword{datasets}
