% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MegaDrought_stack}
\alias{MegaDrought_stack}
\title{Mega drought in Central Chile}
\format{
RasterStack EPSG:32719
}
\source{
\url{https://lpdaac.usgs.gov/}
}
\usage{
MegaDrought_stack
}
\description{
MODIS NDVI RasterStack of Central Chile combining MOD13Q1 & MYD13Q1 between 2000.02.18 and 2021.06.26, 
cleaned using the detailed QA band. Megadrought affected Central Chile between 2010 and 2015 and again in 2019-2022
}
\keyword{datasets}
