% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{modis_dates}
\alias{modis_dates}
\title{A vector with dates for both RasterStack datasets: Mega drought and Blooming Desert, with 929 dates}
\format{
Vector of dates
}
\source{
\url{https://lpdaac.usgs.gov/}
}
\usage{
modis_dates
}
\description{
A vector with dates for both RasterStack datasets: Mega drought and Blooming Desert, with 929 dates
}
\keyword{datasets}
