% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phen.R
\encoding{UTF-8}
\name{Phen}
\alias{Phen}
\title{Phen}
\usage{
Phen(x, dates, h, nGS, rge)
}
\arguments{
\item{x}{Numeric vector with greenness values}

\item{dates}{Vector with dates at which the greenness values were recorded}

\item{h}{Numeric. Indicates the geographic hemisphere to define the starting date of the growing season. h=1 if the vegetation is in the Northern Hemisphere (season starting at January 1st), h=2 if it is in the Southern Hemisphere (season starting at July 1st)}

\item{nGS}{Numeric. Number of greenness values within a single growing season. For example, nGS=23 for MODIS    Vegetation Index 16-days composites}

\item{rge}{A vector containing minimum and maximum values of the response variable used in the analysis. We suggest the use of theoretically based limits. For example in the case of MODIS NDVI or EVI, it ranges from 0 to 10,000, so rge =c(0,10000)}
}
\value{
A numeric vector of length = nGS, where each value represents the expected greeness at that date
}
\description{
Estimates the annual phenological cycle from a time series of vegetation greenness.
}
\details{
Derives the annual phenological cycle for a standard growing season using a numeric vector of vegetation canopy greenness values (e.g. Leaf Area Index, LAI) or satellite based greenness proxies such as the Normalized Difference Vegetation Index (NDVI) or Enhanced Vegetation Index (EVI). A vector with dates for the greenness values is also required.
}
\examples{
\dontshow{
## Testing function with time series of Slovenian data (EVI)
# Load data
phents<-read.table(system.file("extdata/date_tables/datats",package="npphen"),
dec='.',sep='\t',header=TRUE)
# Phenology for the given data
Phen(x=as.vector(phents$x),dates=phents$dates,h=1,nGS=23,rge=c(0,10000))
}
\donttest{
library(lubridate)

## Testing North Hemisphere data. Raster data from Slovenia (EVI index), h=1 ##

# Load data
sl.path<-system.file("extdata/HN_slovenia",package="npphen")
sl_rasters<-list.files(path=sl.path, pattern=glob2rx("slovenia*.tif"), full.names=TRUE)
Slovenia_rasters<-stack(sl_rasters)
sl_dates<-read.csv(system.file("extdata/date_tables/Slovenia_dates.csv", package="npphen"))
Slovenia_dates <- as.Date(sl_dates$date, format='\%d/\%m/\%Y')

# Generate a Raster time series using a raster stack and a date database from Slovenia
# Obtain data from a particular pixel generating a time series
sl_pixel<-cellFromXY(Slovenia_rasters,c(474368,5096979))
sl_pixelts<-as.numeric(Slovenia_rasters[sl_pixel])
plot(Slovenia_dates,sl_pixelts, type='l')

# Phenology for the given pixel
Phen(x=sl_pixelts,dates=Slovenia_dates,h=1,nGS=23,rge=c(0,10000))


## Testing South Hemisphere data. Raster data from Chile (EVI index), h=2 ##

# Load data
ay.path<-system.file("extdata/HS_aysen",package="npphen")
ayrasters<-list.files(path=ay.path, pattern=glob2rx("aysen*.tif"), full.names=TRUE)
Aysen_rasters<-stack(ayrasters)
ay_dates<-read.csv(system.file("extdata/date_tables/Aysen_dates.csv", package="npphen"))
Aysen_dates <- as.Date(ay_dates$date, format='\%d/\%m/\%Y')

# Generate a Raster time series using a raster stack and a date database from Aysen
# Obtain data from a particular pixel generating a time series
ay_pixel<-cellFromXY(Aysen_rasters,c(228373,4806975))
ay_pixelts<-as.numeric(Aysen_rasters[ay_pixel])
plot(Aysen_dates,ay_pixelts, type = 'l')

# Phenology for the given pixel
Phen(x=ay_pixelts,dates=Aysen_dates,h=2,nGS=23,rge=c(0,10000))
}
}
\seealso{
\code{\link{PhenMap}}
}
