% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phents}
\alias{phents}
\title{A dataframe with 8-days MODIS NDVI of a pixel with deciduous Nothofagus macrocarpa forest, Central Chile,
using MOD13Q1/MYD13Q1 combined}
\format{
A data frame with 929 rows and 2 variables:
\describe{
  \item{dates}{dates for each record}
  \item{NDVI}{NDVI value for a pixel of deciduous N. macrocarpa forest}
}
}
\source{
\url{https://lpdaac.usgs.gov/}
}
\usage{
phents
}
\description{
A dataframe with 8-days MODIS NDVI of a pixel with deciduous Nothofagus macrocarpa forest, Central Chile,
using MOD13Q1/MYD13Q1 combined
}
\keyword{datasets}
